+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1]
+
+### Fixed
+
+- Fixed README.md
+
+### Added
+
+- Components can have an external id, which represents exactly the components id specified in a model file (#4)
+- External ids will be used for better error messages for easier finding a specific erroneous component in a 
+  model file (#2)
+
 ## [1.1.0]
 
 ### Changed
@@ -14,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The model registry can be configured to load the latest available model version, if the requested version is not
   known. If no appropriate language can be found, english will be chosen. (#8)
 - In relaxed mode latest available model version loading will be active (#8)
+- Added model database file for 1.5 models (#12)
+- Added new date_time data type (#20)
 - Updated all thirdparty components (#7)
 - Add changelog to the package (#16)